fix multi request queue behaviour, connect monitoring to the request execution pipeline@

+++ logos/db/init.sql
@@ -192,7 +192,7 @@ CREATE TABLE usage_tokens (
     id SERIAL PRIMARY KEY,
     type_id INTEGER NOT NULL REFERENCES token_types(id) ON DELETE CASCADE,
     log_entry_id INTEGER NOT NULL REFERENCES log_entry(id) ON DELETE CASCADE,
-    token_count INTEGER DEFAULT(0)
+    token_count BIGINT DEFAULT(0)
 );
 
 CREATE TABLE token_prices (
